+++ src/map_generator/noise/gradient.rs
@@ -1,4 +1,4 @@
-use core::f32;
+use core::f64;
 
 use super::Noise;
 use crate::map_generator::Interpolation;
@@ -10,12 +10,12 @@ use rand::{
 
 #[derive(Debug, Clone, Copy)]
 struct Vector {
-    x: f32,
-    y: f32,
+    x: f64,
+    y: f64,
 }
 
 impl Vector {
-    fn new(x: f32, y: f32) -> Self {
+    fn new(x: f64, y: f64) -> Self {
         Self { x, y }
     }
 
@@ -23,7 +23,7 @@ impl Vector {
         Self::new(0.0, 0.0)
     }
 
-    fn dot(&self, other: Self) -> f32 {
+    fn dot(&self, other: Self) -> f64 {
         self.x * other.x + self.y * other.y
     }
 }
@@ -31,7 +31,7 @@ impl Vector {
 pub struct Gradient {
     permutation: [u8; 256],
     gradients: [Vector; 256],
-    interpolation: fn(f32, f32, f32) -> f32,
+    interpolation: fn(f64, f64, f64) -> f64,
 }
 
 impl Gradient {
@@ -42,7 +42,7 @@ impl Gradient {
         }
         permutation.shuffle(rng);
 
-        let distribution = Uniform::from(0.0..f32::consts::PI * 2.0);
+        let distribution = Uniform::from(0.0..f64::consts::PI * 2.0);
         let mut gradients = [Vector::zero(); 256];
         for g in gradients.iter_mut() {
             let (s, c) = distribution.sample(rng).sin_cos();
@@ -62,7 +62,7 @@ impl Gradient {
 }
 
 impl Noise for Gradient {
-    fn noise(&self, x: f32, y: f32) -> f32 {
+    fn noise(&self, x: f64, y: f64) -> f64 {
         let xint = x as usize;
         let yint = y as usize;
 
@@ -72,11 +72,12 @@ impl Noise for Gradient {
         let nw = self.gradients[self.index(xint, yint)].dot(Vector::new(xf, yf));
         let ne = self.gradients[self.index(xint + 1, yint)].dot(Vector::new(xf - 1.0, yf));
         let sw = self.gradients[self.index(xint, yint + 1)].dot(Vector::new(xf, yf - 1.0));
-        let se = self.gradients[self.index(xint + 1, yint + 1)].dot(Vector::new(xf - 1.0, yf - 1.0));
+        let se =
+            self.gradients[self.index(xint + 1, yint + 1)].dot(Vector::new(xf - 1.0, yf - 1.0));
 
         let n = (self.interpolation)(nw, ne, xf);
         let s = (self.interpolation)(sw, se, xf);
 
-        (self.interpolation)(n, s, yf) / f32::consts::SQRT_2 + 0.5
+        (self.interpolation)(n, s, yf) / f64::consts::SQRT_2 + 0.5
     }
 }
